Fix #13: use the correct renderer for "me" mob.

+++ src/main/java/noobanidus/mods/miniatures/client/renderer/layers/ArrowRenderTypeLayer.java
@@ -5,10 +5,10 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.EntityRenderDispatcher;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
+import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.projectile.Arrow;
-import net.minecraft.util.Mth;
 import noobanidus.mods.miniatures.client.model.PlayerRenderModel;
 
 // TODO:?
@@ -27,8 +27,8 @@ public class ArrowRenderTypeLayer<T extends LivingEntity, M extends PlayerRender
   protected void renderStuckItem(PoseStack p_225632_1_, MultiBufferSource p_225632_2_, int p_225632_3_, Entity p_225632_4_, float p_225632_5_, float p_225632_6_, float p_225632_7_, float p_225632_8_) {
     float f = Mth.sqrt(p_225632_5_ * p_225632_5_ + p_225632_7_ * p_225632_7_);
     Arrow arrow = new Arrow(p_225632_4_.level, p_225632_4_.getX(), p_225632_4_.getY(), p_225632_4_.getZ());
-    arrow.setYRot((float) (Math.atan2((double) p_225632_5_, (double) p_225632_7_) * (double) (180F / (float) Math.PI)));
-    arrow.setXRot((float) (Math.atan2((double) p_225632_6_, (double) f) * (double) (180F / (float) Math.PI)));
+    arrow.setYRot((float) (Math.atan2(p_225632_5_, p_225632_7_) * (double) (180F / (float) Math.PI)));
+    arrow.setXRot((float) (Math.atan2(p_225632_6_, f) * (double) (180F / (float) Math.PI)));
     arrow.yRotO = arrow.getYRot();
     arrow.xRotO = arrow.getXRot();
     this.dispatcher.render(arrow, 0.0D, 0.0D, 0.0D, 0.0F, p_225632_8_, p_225632_1_, p_225632_2_, p_225632_3_);
