+++ src/main/java/noobanidus/mods/miniatures/client/renderer/layers/BeeStingerRenderTypeLayer.java
@@ -2,17 +2,17 @@ package noobanidus.mods.miniatures.client.renderer.layers;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.mojang.math.Matrix3f;
+import com.mojang.math.Matrix4f;
+import com.mojang.math.Vector3f;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.client.renderer.texture.OverlayTexture;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
-import com.mojang.math.Matrix3f;
-import com.mojang.math.Matrix4f;
-import com.mojang.math.Vector3f;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 import noobanidus.mods.miniatures.client.model.PlayerRenderModel;
 
 public class BeeStingerRenderTypeLayer<T extends LivingEntity, M extends PlayerRenderModel<T>> extends StuckInBodyRenderTypeLayer<T, M> {
@@ -28,8 +28,8 @@ public class BeeStingerRenderTypeLayer<T extends LivingEntity, M extends PlayerR
 
   protected void renderStuckItem(PoseStack p_225632_1_, MultiBufferSource p_225632_2_, int p_225632_3_, Entity p_225632_4_, float p_225632_5_, float p_225632_6_, float p_225632_7_, float p_225632_8_) {
     float f = Mth.sqrt(p_225632_5_ * p_225632_5_ + p_225632_7_ * p_225632_7_);
-    float f1 = (float) (Math.atan2((double) p_225632_5_, (double) p_225632_7_) * (double) (180F / (float) Math.PI));
-    float f2 = (float) (Math.atan2((double) p_225632_6_, (double) f) * (double) (180F / (float) Math.PI));
+    float f1 = (float) (Math.atan2(p_225632_5_, p_225632_7_) * (double) (180F / (float) Math.PI));
+    float f2 = (float) (Math.atan2(p_225632_6_, f) * (double) (180F / (float) Math.PI));
     p_225632_1_.translate(0.0D, 0.0D, 0.0D);
     p_225632_1_.mulPose(Vector3f.YP.rotationDegrees(f1 - 90.0F));
     p_225632_1_.mulPose(Vector3f.ZP.rotationDegrees(f2));
