+++ src/main/java/noobanidus/mods/miniatures/client/renderer/layers/StuckInBodyRenderTypeLayer.java
@@ -1,13 +1,13 @@
 package noobanidus.mods.miniatures.client.renderer.layers;
 
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.client.renderer.entity.layers.RenderLayer;
-import net.minecraft.client.model.geom.ModelPart;
+import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.util.Mth;
 import noobanidus.mods.miniatures.client.model.PlayerRenderModel;
 
 import java.util.Random;
@@ -23,7 +23,7 @@ public abstract class StuckInBodyRenderTypeLayer<T extends LivingEntity, M exten
 
   public void render(PoseStack poseStack, MultiBufferSource bufferIn, int packedLightIn, T entitylivingbaseIn, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch) {
     int i = this.numStuck(entitylivingbaseIn);
-    Random random = new Random((long) entitylivingbaseIn.getId());
+    Random random = new Random(entitylivingbaseIn.getId());
     if (i > 0) {
       for (int j = 0; j < i; ++j) {
         poseStack.pushPose();
@@ -36,7 +36,7 @@ public abstract class StuckInBodyRenderTypeLayer<T extends LivingEntity, M exten
         float f3 = Mth.lerp(f, modelpart$cube.minX, modelpart$cube.maxX) / 16.0F;
         float f4 = Mth.lerp(f1, modelpart$cube.minY, modelpart$cube.maxY) / 16.0F;
         float f5 = Mth.lerp(f2, modelpart$cube.minZ, modelpart$cube.maxZ) / 16.0F;
-        poseStack.translate((double) f3, (double) f4, (double) f5);
+        poseStack.translate(f3, f4, f5);
         f = -1.0F * (f * 2.0F - 1.0F);
         f1 = -1.0F * (f1 * 2.0F - 1.0F);
         f2 = -1.0F * (f2 * 2.0F - 1.0F);
